restructuring files, and deleting others in an attempt to refactor the code
