+++ src/main/webapp/app/app.tsx
@@ -1,5 +1,5 @@
 import 'react-toastify/dist/ReactToastify.css';
-import './app.scss';
+import '../app/styles/app.scss';
 
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
@@ -36,6 +36,10 @@ Sentry.init({
   // for finer control
   tracesSampleRate: 1.0,
 });
+
+// Make sure to call `loadStripe` outside of a component’s render to avoid
+// recreating the `Stripe` object on every render.
+const stripePromise = loadStripe('pk_test_JJ1eMdKN0Hp4UFJ6kWXWO4ix00jtXzq5XG');
 
 export const App = (props: IAppProps) => {
   useEffect(() => {
