+++ src/main/webapp/app/assets/bounties-jdl.png
