+++ src/main/webapp/app/components/bounty/BountyDetails.tsx
