+++ src/main/webapp/app/components/bounty/BountyStepper.tsx
