+++ src/main/webapp/app/components/bounty/bounty-timer/bounty-timer.tsx
