--- src/main/webapp/app/components/landing-page-component.tsx
@@ -1,241 +1,0 @@
-// import './home.scss';
-
-import React from 'react';
-
-import { Segment, Container, Image, Header, Button, Divider, Grid, List } from 'semantic-ui-react';
-import { DesktopContainer, MobileHeader } from 'app/modules/home/home-header';
-import { createMedia } from '@artsy/fresnel';
-import {CardElement, useStripe, useElements} from '@stripe/react-stripe-js';
-export interface ILandingPageComponentProps {
-    account: any,
-    isAuthenticated: any,
-    createPaymentIntent: any,
-};
-
-const { MediaContextProvider, Media } = createMedia({
-	breakpoints: {
-		mobile: 0,
-		tablet: 768,
-		computer: 1024,
-	},
-})
-
-const CARD_OPTIONS = {
-  // iconStyle: 'solid',
-  style: {
-    base: {
-      iconColor: '#c4f0ff',
-      color: '#fff',
-      fontWeight: "500",
-      fontFamily: 'Roboto, Open Sans, Segoe UI, sans-serif',
-      fontSize: '16px',
-      fontSmoothing: 'antialiased',
-      ':-webkit-autofill': {color: '#fce883'},
-      '::placeholder': {color: '#87bbfd'},
-    },
-    invalid: {
-      iconColor: '#ffc7ee',
-      color: '#ffc7ee',
-    },
-  },
-};
-
-export const LandingPageComponent = (props: ILandingPageComponentProps) => {
-  const stripe = useStripe();
-  const elements = useElements();
-
-  const handleSubmit = async (event) => {
-    // Block native form submission.
-    event.preventDefault();
-
-    if (!stripe || !elements) {
-      // Stripe.js has not loaded yet. Make sure to disable
-      // form submission until Stripe.js has loaded.
-      return;
-    }
-
-    // Get a reference to a mounted CardElement. Elements knows how
-    // to find your CardElement because there can only ever be one of
-    // each type of element.
-    const cardElement = elements.getElement(CardElement);
-
-    // Use your card Element with other Stripe.js APIs
-    const {error, paymentMethod} = await stripe.createPaymentMethod({
-      type: 'card',
-      card: cardElement,
-    });
-
-    if (error) {
-      // eslint-disable-next-line no-console
-      console.log('[error]', error);
-    } else {
-      // eslint-disable-next-line no-console
-      console.log('[PaymentMethod]', paymentMethod);
-    }
-  };
-
-  const handleStripePay = () => {
-    props.createPaymentIntent(1);
-  }
-  
-  return (
-    <div>
-      <MediaContextProvider>
-        <Media greaterThan='mobile'>
-          <DesktopContainer
-            // isAdmin={props.isAdmin}
-            isAuthenticated={props.isAuthenticated}
-          />
-        </Media>
-        <Media at='mobile'>
-          <MobileHeader
-          // isAdmin={props.isAdmin}
-          isAuthenticated={props.isAuthenticated}
-        />
-        </Media>
-      </MediaContextProvider>
-      
-      <Segment style={{ padding: '8em 0em' }} vertical>
-        <Grid container stackable verticalAlign='middle'>
-          <Grid.Row>
-            <Grid.Column width={8}>
-              <Header as='h3' style={{ fontSize: '2em' }}>
-                We Help Companies and Companions
-              </Header>
-              
-              <Button onClick={handleStripePay}>Stripe Pay</Button>
-
-              {/* <form onSubmit={handleSubmit}>
-                <fieldset className="FormGroup">
-                  <div className="FormRow">
-                    <CardElement options={CARD_OPTIONS} />
-                  </div>
-                </fieldset>
-                <button type="submit" disabled={!stripe}>
-                  Pay
-                </button>
-              </form> */}
-
-              <p style={{ fontSize: '1.33em' }}>
-                We can give your company superpowers to do things that they never thought possible.
-                Let us delight your customers and empower your needs... through pure data analytics.
-              </p>
-              <Header as='h3' style={{ fontSize: '2em' }}>
-                We Make Bananas That Can Dance
-              </Header>
-              <p style={{ fontSize: '1.33em' }}>
-                Yes that&apos;s right, you thought it was the stuff of dreams, but even bananas can be
-                bioengineered.
-              </p>
-            </Grid.Column>
-            <Grid.Column floated='right' width={6}>
-              <Image bordered rounded size='large' src='/images/wireframe/white-image.png' />
-            </Grid.Column>
-          </Grid.Row>
-          <Grid.Row>
-            <Grid.Column textAlign='center'>
-              <Button size='huge'>Check Them Out</Button>
-            </Grid.Column>
-          </Grid.Row>
-        </Grid>
-      </Segment>
-
-      <Segment style={{ padding: '0em' }} vertical>
-        <Grid celled='internally' columns='equal' stackable>
-          <Grid.Row textAlign='center'>
-            <Grid.Column style={{ paddingBottom: '5em', paddingTop: '5em' }}>
-              <Header as='h3' style={{ fontSize: '2em' }}>
-                &lsquo;What a Company&rsquo;
-              </Header>
-              <p style={{ fontSize: '1.33em' }}>That is what they all say about us</p>
-            </Grid.Column>
-            <Grid.Column style={{ paddingBottom: '5em', paddingTop: '5em' }}>
-              <Header as='h3' style={{ fontSize: '2em' }}>
-                &lsquo;I shouldn&apos;t have gone with their competitor.&rsquo;
-              </Header>
-              <p style={{ fontSize: '1.33em' }}>
-                <Image avatar src='/images/avatar/large/nan.jpg' />
-                <b>Nan</b> Chief Fun Officer Acme Toys
-              </p>
-            </Grid.Column>
-          </Grid.Row>
-        </Grid>
-      </Segment>
-
-      <Segment style={{ padding: '8em 0em' }} vertical>
-        <Container text>
-          <Header as='h3' style={{ fontSize: '2em' }}>
-            Breaking The Grid, Grabs Your Attention
-          </Header>
-          <p style={{ fontSize: '1.33em' }}>
-            Instead of focusing on content creation and hard work, we have learned how to master the
-            art of doing nothing by providing massive amounts of whitespace and generic content that
-            can seem massive, monolithic and worth your attention.
-          </p>
-          <Button as='a' size='large'>
-            Read More
-          </Button>
-
-          <Divider
-            as='h4'
-            className='header'
-            horizontal
-            style={{ margin: '3em 0em', textTransform: 'uppercase' }}
-          >
-            <a href='#'>Case Studies</a>
-          </Divider>
-
-          <Header as='h3' style={{ fontSize: '2em' }}>
-            Did We Tell You About Our Bananas?
-          </Header>
-          <p style={{ fontSize: '1.33em' }}>
-            Yes I know you probably disregarded the earlier boasts as non-sequitur filler content, but
-            it&apos;s really true. It took years of gene splicing and combinatory DNA research, but our
-            bananas can really dance.
-          </p>
-          <Button as='a' size='large'>
-            I&apos;m Still Quite Interested
-          </Button>
-        </Container>
-      </Segment>
-
-      <Segment inverted vertical style={{ padding: '5em 0em' }}>
-        <Container>
-          <Grid divided inverted stackable>
-            <Grid.Row>
-              <Grid.Column width={3}>
-                <Header inverted as='h4' content='About' />
-                <List link inverted>
-                  <List.Item as='a'>Sitemap</List.Item>
-                  <List.Item as='a'>Contact Us</List.Item>
-                  <List.Item as='a'>Religious Ceremonies</List.Item>
-                  <List.Item as='a'>Gazebo Plans</List.Item>
-                </List>
-              </Grid.Column>
-              <Grid.Column width={3}>
-                <Header inverted as='h4' content='Services' />
-                <List link inverted>
-                  <List.Item as='a'>Bonono Pre-Order</List.Item>
-                  <List.Item as='a'>DNA FAQ</List.Item>
-                  <List.Item as='a'>How To Access</List.Item>
-                  <List.Item as='a'>Favorite X-Men</List.Item>
-                </List>
-              </Grid.Column>
-              <Grid.Column width={7}>
-                <Header as='h4' inverted>
-                  Footer Header
-                </Header>
-                <p>
-                  Extra space for a call to action inside the footer that could help re-engage users.
-                </p>
-              </Grid.Column>
-            </Grid.Row>
-          </Grid>
-        </Container>
-      </Segment>
-  
-    </div>
-  );
-};
-
-export default LandingPageComponent
