+++ src/main/webapp/app/styles/app.scss
@@ -4,7 +4,7 @@
 // @import 'node_modules/bootstrap/scss/bootstrap';
 // @import '~bootswatch/dist/materia/bootswatch';
 .bg-image {
-  background-image: url('./bounties_bg.jpg');
+  background-image: url('../assets/bounties_bg.jpg');
   background-repeat: no-repeat;
   background-size: cover;
   min-height: calc(100vh - 125px);
