+++ src/main/webapp/app/styles/rtl.scss
