+++ src/main/webapp/index.html
@@ -7,7 +7,7 @@
     <meta name="description" content="Description for bounties">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="theme-color" content="#000000">
-    <link rel="icon" href="favicon.ico" />
+    <link rel="icon" href="./app/assets/favicon@2x.ico" />
     <link rel="manifest" href="manifest.webapp" />
     <link rel="stylesheet" href="content/css/loading.css">
     <!-- jhipster-needle-add-resources-to-root - JHipster will add new resources here -->
